+++ modules/api/src/main/scala/vinyldns/api/domain/auth/AuthPrincipalProvider.scala
@@ -20,7 +20,7 @@ import cats.effect._
 import vinyldns.core.domain.membership.{MembershipRepository, User, UserRepository}
 import vinyldns.api.repository.dynamodb.{DynamoDBMembershipRepository, DynamoDBUserRepository}
 import vinyldns.core.domain.auth.AuthPrincipal
-import vinyldns.core.route.Monitored
+import vinyldns.api.route.Monitored
 
 trait AuthPrincipalProvider extends Monitored {
   def getAuthPrincipal(accessKey: String): IO[Option[AuthPrincipal]]
