+++ modules/api/src/main/scala/vinyldns/api/repository/dynamodb/DynamoDBUserRepository.scala
@@ -27,7 +27,7 @@ import org.joda.time.DateTime
 import org.slf4j.{Logger, LoggerFactory}
 import vinyldns.api.VinylDNSConfig
 import vinyldns.core.domain.membership.{ListUsersResults, User, UserRepository}
-import vinyldns.core.route.Monitored
+import vinyldns.api.route.Monitored
 
 import scala.collection.JavaConverters._
