+++ modules/api/src/main/scala/vinyldns/api/route/VinylDNSDirectives.scala
@@ -26,7 +26,6 @@ import cats.effect._
 import org.json4s.JsonDSL._
 import org.json4s.jackson.JsonMethods._
 import vinyldns.core.domain.auth.AuthPrincipal
-import vinyldns.core.route.Monitor
 
 import scala.concurrent.duration._
 import scala.util.Failure
