+++ modules/api/src/test/scala/vinyldns/api/route/VinylDNSDirectivesSpec.scala
@@ -26,7 +26,6 @@ import org.mockito.Matchers._
 import org.mockito.Mockito._
 import org.scalatest.mockito.MockitoSugar
 import org.scalatest.{BeforeAndAfterEach, Matchers, OneInstancePerTest, WordSpec}
-import vinyldns.core.route.Monitor
 
 class VinylDNSDirectivesSpec
     extends WordSpec
