+++ project/Dependencies.scala
@@ -15,6 +15,12 @@ object Dependencies {
   lazy val awsV = "1.11.95"
 
   lazy val compileDependencies = Seq(
+    "com.typesafe.akka"         %% "akka-http"                      % akkaHttpV,
+    "com.typesafe.akka"         %% "akka-http-spray-json"           % akkaHttpV,
+    "de.heikoseeberger"         %% "akka-http-json4s"               % "1.21.0",
+    "com.typesafe.akka"         %% "akka-remote"                    % akkaV,
+    "com.typesafe.akka"         %% "akka-slf4j"                     % akkaV,
+    "com.typesafe.akka"         %% "akka-actor"                     % akkaV,
     "ch.qos.logback"            %  "logback-classic"                % "1.0.7",
     "com.aaronbedra"            %  "orchard"                        % "0.1.1",
     "com.amazonaws"             %  "aws-java-sdk-core"              % awsV withSources(),
@@ -26,6 +32,7 @@ object Dependencies {
     "com.zaxxer"                %  "HikariCP"                       % "2.5.1",
     "dnsjava"                   %  "dnsjava"                        % "2.1.7",
     "org.mariadb.jdbc"          %  "mariadb-java-client"            % "2.2.3",
+    "nl.grons"                  %% "metrics-scala"                  % metricsScalaV,
     "org.apache.commons"        %  "commons-lang3"                  % "3.4",
     "org.flywaydb"              %  "flyway-core"                    % "5.1.4",
     "org.json4s"                %% "json4s-ext"                     % "3.5.3",
@@ -33,6 +40,7 @@ object Dependencies {
     "org.scalikejdbc"           %% "scalikejdbc"                    % "2.5.2",
     "org.scalikejdbc"           %% "scalikejdbc-config"             % "2.5.2",
     "org.scodec"                %% "scodec-bits"                    % scodecV,
+    "org.slf4j"                 %  "slf4j-api"                      % "1.7.7",
     "co.fs2"                    %% "fs2-core"                       % "0.10.5",
     "com.github.pureconfig"     %% "pureconfig"                     % pureConfigV,
     "com.github.pureconfig"     %% "pureconfig-cats-effect"         % pureConfigV,
@@ -47,25 +55,20 @@ object Dependencies {
   lazy val coreDependencies = Seq(
     "org.typelevel"             %% "cats-effect"                    % catsEffectV,
     "com.typesafe"              %  "config"                         % configV,
-    "com.typesafe.akka"         %% "akka-http"                      % akkaHttpV,
-    "com.typesafe.akka"         %% "akka-http-spray-json"           % akkaHttpV,
-    "de.heikoseeberger"         %% "akka-http-json4s"               % "1.21.0",
-    "com.typesafe.akka"         %% "akka-remote"                    % akkaV,
-    "com.typesafe.akka"         %% "akka-slf4j"                     % akkaV,
-    "com.typesafe.akka"         %% "akka-actor"                     % akkaV,
     "joda-time"                 %  "joda-time"                      % "2.8.1",
-    "nl.grons"                  %% "metrics-scala"                  % metricsScalaV,
-    "org.slf4j"                 %  "slf4j-api"                      % "1.7.7",
     "org.scodec"                %% "scodec-bits"                    % scodecV
   )
 
-  lazy val testDependencies = Seq(
-    "com.typesafe.akka"         %% "akka-http-testkit"              % akkaHttpV,
-    "junit"                     %  "junit"                          % "4.12",
-    "org.mockito"               %  "mockito-core"                   % "1.10.19",
+  lazy val coreTestDependencies = Seq(
     "org.scalatest"             %% "scalatest"                      % scalaTestV,
     "org.scalacheck"            %% "scalacheck"                     % "1.13.4",
     "com.ironcorelabs"          %% "cats-scalatest"                 % "2.3.1"
+  )
+
+  lazy val testDependencies = coreTestDependencies ++ Seq(
+    "com.typesafe.akka"         %% "akka-http-testkit"              % akkaHttpV,
+    "junit"                     %  "junit"                          % "4.12",
+    "org.mockito"               %  "mockito-core"                   % "1.10.19"
   )
 
   lazy val portalDependencies = Seq(
